+++ app/routers/employee_routers.py
@@ -4,7 +4,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from entities.users.models_user import EmployeeModel
 from routers.consts import RouteSlug
 from models.database import get_session
-from schemas.user_schemas import EmployeeInSchema, EmployeeInOptionalSchema
+from entities.users.schemas_users import EmployeeInSchema, EmployeeInOptionalSchema
 from services.employee_service import EmployeeService
 from dependencies.base_dependency import (
     ValidGetByIdDependency,
