+++ app/tests/test_employee.py
@@ -2,7 +2,7 @@ from typing import Type
 
 from entities.users.models_user import EmployeeModel
 from tests.abstract_user import UserAbstract
-from schemas.user_schemas import EmployeeInSchema, EmployeeInOptionalSchema
+from entities.users.schemas_users import EmployeeInSchema, EmployeeInOptionalSchema
 from services.stub_init_service import LAST_NAMES
